Enforced complex128 type to FFT in dottest tutorial

+++ tutorials/dottest.py
@@ -129,6 +129,7 @@ dt = 0.005
 nt = 100
 nfft = 2**10
 
-FFTop = pylops.signalprocessing.FFT(dims=(nt,), nfft=nfft, sampling=dt)
+FFTop = pylops.signalprocessing.FFT(dims=(nt,), nfft=nfft,
+                                    sampling=dt, dtype=np.complex128)
 dottest(FFTop, nfft, nt, complexflag=2, verb=True)
 dottest(FFTop, nfft, nt, complexflag=3, verb=True)
